+++ megalinter/linters/PowershellLinter.py
@@ -4,6 +4,7 @@ Use PowerShell to lint Powershell files
 https://github.com/PowerShell/PSScriptAnalyzer
 """
 import sys
+from shutil import get_terminal_size
 
 from megalinter import Linter, config
 
@@ -19,7 +20,11 @@ class PowershellLinter(Linter):
     def build_lint_command(self, file=None):
         pwsh_script = []
         if self.linter_name == "powershell":
-            pwsh_script = ["Invoke-ScriptAnalyzer -EnableExit"]
+            # Prevent Out-String (piped later) to strip ANSI escape sequences
+            # and start forming a call to Invoke-ScriptAnalyzer
+            pwsh_script = [
+                "$PSStyle.OutputRendering = 'Ansi'; Invoke-ScriptAnalyzer -EnableExit"
+            ]
         elif self.linter_name == "powershell_formatter":
             pwsh_script = ["Invoke-Formatter"]
 
@@ -51,6 +56,15 @@ class PowershellLinter(Linter):
             and self.cli_lint_fix_arg_name is not None
         ):
             pwsh_script[0] += f" {self.cli_lint_fix_arg_name}"
+        if self.linter_name == "powershell":
+            # Enforce table output for easier parsing of number of errors
+            # (severity first)
+            pwsh_script[0] += " | Format-Table -AutoSize -Wrap -Property Severity, RuleName, ScriptName, Line, Message"
+            # Format output to fit in terminal, respecting the terminal width.
+            # Use good defaults, shutil respects COLUMNS env var. For more info:
+            # https://stackoverflow.com/a/76889369
+            size = get_terminal_size()
+            pwsh_script[0] += f" | Out-String -Width {size.columns}"
         cmd = [
             *self.cli_executable,
             "-NoProfile",